+++ src/payable/find-by-id/find-payable-by-id.controller.ts
@@ -2,7 +2,7 @@ import { Controller, Get, Param, ParseUUIDPipe } from "@nestjs/common";
 import type { Payable } from "@prisma/client";
 
 import { PayableDTO } from "../payable.dto";
-import { FindPayableByIdPipe } from "./find-payable-by-id.pipe";
+import { FindPayableByIdPipe } from "../find-payable-by-id.pipe";
 
 @Controller()
 export class FindPayableByIdController {
