+++ src/payable/update-by-id/update-payable-by-id.controller.ts
@@ -9,7 +9,7 @@ import {
 import type { Payable } from "@prisma/client";
 
 import { PrismaProvider } from "../../providers/prisma.provider";
-import { FindPayableByIdPipe } from "../find-by-id/find-payable-by-id.pipe";
+import { FindPayableByIdPipe } from "../find-payable-by-id.pipe";
 import { PayableDTO } from "../payable.dto";
 import { UpdatePayableByIdInputDTO } from "./update-payable-by-id-input.dto";
 import { UpdatePayableByIdInputPipe } from "./update-payable-by-id-input.pipe";
